requirements.txt
fastapi
uvicorn[standard]
openai
websocket-client